[MODIFY] Implement Payments feature: add PaymentsContext, PaymentModal, and StudentPaymentsModal; refactor Payments component for global payment handling and improved user experience
